requirements.txt
# FastAPI and Uvicorn for API development
fastapi
uvicorn[standard]

# SQLAlchemy for ORM and database interaction
sqlalchemy

# pandas for data manipulation and preparation
pandas

# python-dotenv for environment variable management
python-dotenv

# matplotlib for visualizations in Quarto reports
matplotlib

# Quarto for generating reports
quartodoc

# pyodbc for connecting to SQL Server or other ODBC-supported databases
pyodbc

# Jupyter for development and testing within notebooks
jupyter

# OpenPyXL for advanced Excel handling
openpyxl

# Requests for making HTTP calls (if needed for fetching data from APIs)
requests

# MarkupSafe for working with Jinja2 templates (used internally by some libraries like Quarto)
MarkupSafe

# Pydantic for schema validation (FastAPI dependency)
pydantic

# Access ODBC_ databases with asyncio
aioodbc

# Pytest and its plugins for testing, mocking, and coverage
pytest-mock

# FastAPI testing utilities
httpx

# Test library to measure code coverage
pytest-cov
